game over enabled
now when a player has played more than 5 rounds the game displays a message
"game over" and the player can't play anymore, Also the score, points, and rounds are now hidden.
 Now when the player hits the
play again button the window refreshes and the game starts over, simplifying the
logic and making the code more simple and readable. for this i used an event listener
on the play again button that calls the function location.reload() which reloads the page.