things look better
i changed the syntax that showed game over, now it has it's own div, which
is hidden until the game is over. I styled all the buttons so the have a little
animation to them=, and the cursor changes to a pointer when you hover over them.
for now i thnk thats it for logic and styling.